FOP-2265: Enable xlint cast; patch submitted by Simon Steiner
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@1502902 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/afp/modca/MapPageOverlay.java
@@ -99,8 +99,8 @@ public class MapPageOverlay extends AbstractAFPObject {
 
         // Structured field ID for a MPO
         data[3] = (byte) 0xD3;
-        data[4] = (byte) Type.MAP;
-        data[5] = (byte) Category.PAGE_OVERLAY;
+        data[4] = Type.MAP;
+        data[5] = Category.PAGE_OVERLAY;
 
         data[6] = 0x00; // Reserved
         data[7] = 0x00; // Reserved
